// src/FE/src/api/generated/models/worksheetResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface WorksheetResponse {
  id: string;
  organizationId: string;
  jobId: string;
  userId: string;
  userDisplayName: string;
  workDate: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  hoursWorked: number | string;
  sleptOnJob: boolean;
  createdAt: string;
  updatedAt: string;
}